# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/memberships": {
      "get": {
        "operationId": "MembershipsController_getMyMemberships",
        "summary": "List all tenants the authenticated user belongs to",
        "tags": ["Memberships"],
        "security": [{ "bearer": [] }],
        "responses": {
          "200": { "description": "Array of memberships with tenant details" }
        }
      },
      "post": {
        "operationId": "MembershipsController_createMembership",
        "summary": "Add a user to the current tenant (admin/pastor only)",
        "tags": ["Memberships"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateMembershipDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Membership created" },
          "404": { "description": "User with given email not found" }
        }
      }
    },
    "/api/tenants/{tenantId}/members": {
      "get": {
        "operationId": "MembershipsController_getMembers",
        "summary": "List tenant members with cursor-based pagination",
        "tags": ["Memberships"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" }, "description": "user_id of last member seen" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "Paginated member list with nextCursor" }
        }
      }
    },
    "/api/tenants/{tenantId}/members/{userId}/role": {
      "patch": {
        "operationId": "MembershipsController_updateRole",
        "summary": "Update a member role (admin only)",
        "tags": ["Memberships"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateRoleDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Role updated" },
          "404": { "description": "Membership not found or not authorized" }
        }
      }
    },
    "/api/tenants/{tenantId}/members/{userId}": {
      "delete": {
        "operationId": "MembershipsController_removeMember",
        "summary": "Remove a member from tenant (admin or self)",
        "tags": ["Memberships"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Member removed" },
          "404": { "description": "Membership not found or not authorized" }
        }
      }
    },
...
    "/api/invitations": {
      "get": {
        "operationId": "InvitationsController_getInvitations",
        "summary": "List pending invitations for current tenant (admin/pastor)",
        "tags": ["Invitations"],
        "security": [{ "bearer": [] }],
        "responses": {
          "200": { "description": "Array of pending invitations" }
        }
      },
      "post": {
        "operationId": "InvitationsController_createInvitation",
        "summary": "Send an invitation to join the current tenant",
        "tags": ["Invitations"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateInvitationDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "Invitation created and email queued" }
        }
      }
    },
    "/api/invitations/{token}/accept": {
      "post": {
        "operationId": "InvitationsController_acceptInvitation",
        "summary": "Accept an invitation by token. Call /auth/switch-tenant + /auth/refresh afterward.",
        "tags": ["Invitations"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "token", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Invitation accepted. Membership created." },
          "400": { "description": "Invalid/expired token or email mismatch" }
        }
      }
    },
    "/api/notifications": {
      "get": {
        "operationId": "NotificationsController_getNotifications",
        "summary": "List notifications for authenticated user",
        "tags": ["Notifications"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 50 } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer", "default": 0, "minimum": 0 } },
          { "name": "unreadOnly", "in": "query", "required": false, "schema": { "type": "boolean", "default": false }, "description": "When true, returns only unread notifications" }
        ],
        "responses": {
          "200": { "description": "Paginated array of notifications" }
        }
      }
    },
    "/api/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markAsRead",
        "summary": "Mark a notification as read",
        "tags": ["Notifications"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Notification marked as read" },
          "404": { "description": "Notification not found" }
        }
      }
    },
    "/api/media/presigned-url": {
      "post": {
        "operationId": "MediaController_generatePresignedUrl",
        "summary": "Generate a pre-signed S3 upload URL (5 min expiry)",
        "tags": ["Media"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PresignedUrlDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "Returns uploadUrl and fileKey for direct S3 upload" },
          "400": { "description": "No active tenant context" }
        }
      }
    },
...
    "/api/giving/donate": {
      "post": {
        "operationId": "GivingController_donate",
        "summary": "Create a donation PaymentIntent (returns clientSecret for Stripe.js)",
        "tags": ["Giving"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DonateDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "{ clientSecret, transactionId }" },
          "400": { "description": "Church has not set up or activated payment processing" }
        }
      }
    },
    "/api/giving/transactions": {
      "get": {
        "operationId": "GivingController_getMyTransactions",
        "summary": "Get authenticated user donation history (cursor-paginated)",
        "tags": ["Giving"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "string", "default": "20" } }
        ],
        "responses": {
          "200": { "description": "{ transactions, nextCursor }" }
        }
      }
    },
    "/api/tenants/{tenantId}/transactions": {
      "get": {
        "operationId": "GivingController_getTenantTransactions",
        "summary": "Get all tenant transactions (admin dashboard, cursor-paginated)",
        "tags": ["Giving"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "tenantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "string", "default": "20" } }
        ],
        "responses": {
          "200": { "description": "{ transactions, nextCursor }" }
        }
      }
    },
...
  },
  "info": {
    "title": "ChurchApp Platform API",
    "description": "Multi-tenant church management platform & global social network.\n\n**Authentication:** All protected endpoints require a Supabase JWT in the `Authorization: Bearer <token>` header.\n\n**Tenant Context:** Most endpoints are tenant-scoped via PostgreSQL Row-Level Security. Call `POST /api/auth/switch-tenant` followed by `POST /api/auth/refresh` to set the active tenant in your JWT.\n\n**Rate Limits:** Global: 100 req/min per IP. Auth endpoints: 5 req/min per IP. Webhooks: unlimited.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase JWT access token"
      }
    },
    "schemas": {
...
      "CreateMembershipDto": {
        "type": "object",
        "required": ["email", "role"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "newmember@example.com", "description": "Email of the user to add. Must already have a platform account." },
          "role": { "type": "string", "enum": ["admin", "pastor", "member"], "example": "member" }
        }
      },
      "UpdateRoleDto": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": { "type": "string", "enum": ["admin", "pastor", "member"], "example": "pastor" }
        }
      },
...
      "CreateInvitationDto": {
        "type": "object",
        "required": ["email", "role"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "invitee@example.com", "description": "Email address of the person being invited" },
          "role": { "type": "string", "enum": ["admin", "pastor", "member"], "example": "member" }
        }
      },
      "PresignedUrlDto": {
        "type": "object",
        "required": ["filename", "contentType"],
        "properties": {
          "filename": { "type": "string", "maxLength": 255, "example": "sermon-photo.jpg" },
          "contentType": { "type": "string", "example": "image/jpeg", "description": "Supported: image/jpeg, image/png, image/gif, image/webp, video/mp4, video/quicktime, video/webm" }
        }
      },
...
      "DonateDto": {
        "type": "object",
        "required": ["amount"],
        "properties": {
          "amount": { "type": "number", "minimum": 1, "example": 100, "description": "Donation amount in dollars (e.g., 100 = $100.00)" },
          "currency": { "type": "string", "enum": ["usd", "eur", "gbp", "cad", "aud"], "default": "usd" }
        }
      }
    }
  }
}